Modify include file

+++ src/Common.c
@@ -1,6 +1,6 @@
 #include <math.h>
-
-float Common_GetNorm(float* var, size_t n) {
+#include "Common.h"
+float Common_GetNorm(float* var, unsigned int n) {
     int i;
     float sum = 0;
     for (i=0; i<n; ++i) sum += var[i]*var[i];
